+++ cloudinary-core/pom.xml
@@ -7,7 +7,7 @@
         <version>1.1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>cloudinary</artifactId>
+    <artifactId>cloudinary-core</artifactId>
     <packaging>jar</packaging>
 
 </project>
