+++ cloudinary-core/src/main/java/com/cloudinary/utils/ObjectUtils.java
@@ -9,9 +9,10 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
+import org.cloudinary.json.JSONArray;
+import org.cloudinary.json.JSONException;
+import org.cloudinary.json.JSONObject;
+
 
 public class ObjectUtils {
