+++ cloudinary-core/src/main/java/org/cloudinary/json/JSONException.java
@@ -1,4 +1,4 @@
-package org.json;
+package org.cloudinary.json;
 
 /**
  * The JSONException is thrown by the JSON.org classes when things are amiss.
