+++ cloudinary-core/src/main/java/org/cloudinary/json/JSONTokener.java
@@ -1,4 +1,4 @@
-package org.json;
+package org.cloudinary.json;
 
 import java.io.BufferedReader;
 import java.io.IOException;
