+++ cloudinary-taglib/pom.xml
@@ -15,7 +15,7 @@
     <dependencies>
         <dependency>
             <groupId>com.cloudinary</groupId>
-            <artifactId>cloudinary</artifactId>
+            <artifactId>cloudinary-http42</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
